# sqlstats: improve perf and memory consumption of json encode/decode