--- a/pkg/sql/sqlstats/persistedsqlstats/sqlstatsutil/BUILD.bazel
+++ b/pkg/sql/sqlstats/persistedsqlstats/sqlstatsutil/BUILD.bazel
@@ -14,6 +14,7 @@ go_library(
     deps = [
         "//pkg/sql/appstatspb",
         "//pkg/sql/sem/tree",
+        "//pkg/util/buildutil",
         "//pkg/util/encoding",
         "//pkg/util/json",
         "//pkg/util/timeutil",
@@ -31,6 +32,7 @@ go_test(
     embed = [":sqlstatsutil"],
     deps = [
         "//pkg/sql/appstatspb",
+        "//pkg/testutils",
         "//pkg/util/json",
         "//pkg/util/leaktest",
         "//pkg/util/log",