--- a/pkg/sql/sqlstats/persistedsqlstats/sqlstatsutil/json_decoding.go
+++ b/pkg/sql/sqlstats/persistedsqlstats/sqlstatsutil/json_decoding.go
@@ -7,9 +7,12 @@ package sqlstatsutil
 
 import (
 	"github.com/cockroachdb/cockroach/pkg/sql/appstatspb"
+	"github.com/cockroachdb/cockroach/pkg/util/buildutil"
 	"github.com/cockroachdb/cockroach/pkg/util/json"
 	"github.com/cockroachdb/errors"
 )
+
+var panicOnUnknownFieldsForTest = true
 
 // DecodeTxnStatsMetadataJSON decodes the 'metadata' field of the JSON
 // representation of transaction statistics into
@@ -17,9 +20,14 @@ import (
 func DecodeTxnStatsMetadataJSON(
 	metadata json.JSON, result *appstatspb.CollectedTransactionStatistics,
 ) error {
-	return jsonFields{
-		{"stmtFingerprintIDs", (*stmtFingerprintIDArray)(&result.StatementFingerprintIDs)},
-	}.decodeJSON(metadata)
+	valJSON, err := metadata.FetchValKey("stmtFingerprintIDs")
+	if err != nil {
+		return err
+	}
+	if valJSON != nil {
+		return (*stmtFingerprintIDArray)(&result.StatementFingerprintIDs).decodeJSON(valJSON)
+	}
+	return nil
 }
 
 // DecodeTxnStatsStatisticsJSON decodes the 'statistics' section of the
@@ -28,7 +36,23 @@ func DecodeTxnStatsMetadataJSON(
 func DecodeTxnStatsStatisticsJSON(
 	jsonVal json.JSON, result *appstatspb.TransactionStatistics,
 ) error {
-	return (*txnStats)(result).decodeJSON(jsonVal)
+	if statsJSON, err := jsonVal.FetchValKey("statistics"); err != nil {
+		return err
+	} else if statsJSON != nil {
+		if err := (*innerTxnStats)(result).decodeJSON(statsJSON); err != nil {
+			return err
+		}
+	}
+
+	if execJSON, err := jsonVal.FetchValKey("execution_statistics"); err != nil {
+		return err
+	} else if execJSON != nil {
+		if err := (*execStats)(&result.ExecStats).decodeJSON(execJSON); err != nil {
+			return err
+		}
+	}
+
+	return nil
 }
 
 // DecodeStmtStatsMetadataJSON decodes the 'metadata' field of the JSON
@@ -37,7 +61,7 @@ func DecodeTxnStatsStatisticsJSON(
 func DecodeStmtStatsMetadataJSON(
 	metadata json.JSON, result *appstatspb.CollectedStatementStatistics,
 ) error {
-	return (*stmtStatsMetadata)(result).jsonFields().decodeJSON(metadata)
+	return (*stmtStatsMetadata)(result).decodeJSON(metadata)
 }
 
 // DecodeStmtStatsMetadataFlagsOnlyJSON decodes the 'metadata' flags only fields
@@ -47,14 +71,274 @@ func DecodeStmtStatsMetadataJSON(
 func DecodeStmtStatsMetadataFlagsOnlyJSON(
 	metadata json.JSON, result *appstatspb.CollectedStatementStatistics,
 ) error {
-	return (*stmtStatsMetadata)(result).jsonFlagsOnlyFields().decodeJSON(metadata)
+	return (*stmtStatsMetadata)(result).decodeFlagsOnlyJSON(metadata)
 }
 
 // DecodeAggregatedMetadataJSON decodes the 'aggregated metadata' represented by appstatspb.AggregatedStatementMetadata.
 func DecodeAggregatedMetadataJSON(
 	metadata json.JSON, result *appstatspb.AggregatedStatementMetadata,
 ) error {
-	return (*aggregatedMetadata)(result).jsonFields().decodeJSON(metadata)
+	return (*aggregatedMetadata)(result).decodeJSON(metadata)
+}
+
+func (am *aggregatedMetadata) decodeJSON(js json.JSON) (err error) {
+	iter, err := js.ObjectIter()
+	if err != nil {
+		return err
+	}
+
+	var fieldName string
+	defer func() {
+		if err != nil {
+			err = errors.Wrapf(err, "decoding field %s", fieldName)
+		}
+	}()
+
+	for ok := iter.Next(); ok; ok = iter.Next() {
+		switch iter.Key() {
+		case "fingerprintID":
+			err = (*jsonString)(&am.FingerprintID).decodeJSON(iter.Value())
+			if err != nil {
+				return err
+			}
+		case "db":
+			err = (*stringArray)(&am.Databases).decodeJSON(iter.Value())
+			if err != nil {
+				return err
+			}
+		case "appNames":
+			err = (*stringArray)(&am.AppNames).decodeJSON(iter.Value())
+			if err != nil {
+				return err
+			}
+		case "distSQLCount":
+			err = (*jsonInt)(&am.DistSQLCount).decodeJSON(iter.Value())
+			if err != nil {
+				return err
+			}
+		case "fullScanCount":
+			err = (*jsonInt)(&am.FullScanCount).decodeJSON(iter.Value())
+			if err != nil {
+				return err
+			}
+		case "implicitTxn":
+			err = (*jsonBool)(&am.ImplicitTxn).decodeJSON(iter.Value())
+			if err != nil {
+				return err
+			}
+		case "query":
+			err = (*jsonString)(&am.Query).decodeJSON(iter.Value())
+			if err != nil {
+				return err
+			}
+		case "formattedQuery":
+			err = (*jsonString)(&am.FormattedQuery).decodeJSON(iter.Value())
+			if err != nil {
+				return err
+			}
+		case "querySummary":
+			err = (*jsonString)(&am.QuerySummary).decodeJSON(iter.Value())
+			if err != nil {
+				return err
+			}
+		case "stmtType":
+			err = (*jsonString)(&am.StmtType).decodeJSON(iter.Value())
+			if err != nil {
+				return err
+			}
+		case "vecCount":
+			err = (*jsonInt)(&am.VecCount).decodeJSON(iter.Value())
+			if err != nil {
+				return err
+			}
+		case "totalCount":
+			err = (*jsonInt)(&am.TotalCount).decodeJSON(iter.Value())
+			if err != nil {
+				return err
+			}
+		default:
+			if buildutil.CrdbTestBuild && panicOnUnknownFieldsForTest {
+				panic(errors.AssertionFailedf("unknown field: %s", iter.Key()))
+			}
+		}
+	}
+
+	return nil
+}
+
+func (sm *stmtStatsMetadata) decodeJSON(js json.JSON) (err error) {
+	iter, err := js.ObjectIter()
+	if err != nil {
+		return err
+	}
+
+	var fieldName string
+	defer func() {
+		if err != nil {
+			err = errors.Wrapf(err, "decoding field %s", fieldName)
+		}
+	}()
+
+	for ok := iter.Next(); ok; ok = iter.Next() {
+		switch iter.Key() {
+		case "stmtType":
+			err = (*jsonString)(&sm.Stats.SQLType).decodeJSON(iter.Value())
+			if err != nil {
+				return err
+			}
+		case "query":
+			err = (*jsonString)(&sm.Key.Query).decodeJSON(iter.Value())
+			if err != nil {
+				return err
+			}
+		case "querySummary":
+			err = (*jsonString)(&sm.Key.QuerySummary).decodeJSON(iter.Value())
+			if err != nil {
+				return err
+			}
+		case "db":
+			err = (*jsonString)(&sm.Key.Database).decodeJSON(iter.Value())
+			if err != nil {
+				return err
+			}
+		case "distsql":
+			err = (*jsonBool)(&sm.Key.DistSQL).decodeJSON(iter.Value())
+			if err != nil {
+				return err
+			}
+		case "implicitTxn":
+			err = (*jsonBool)(&sm.Key.ImplicitTxn).decodeJSON(iter.Value())
+			if err != nil {
+				return err
+			}
+		case "vec":
+			err = (*jsonBool)(&sm.Key.Vec).decodeJSON(iter.Value())
+			if err != nil {
+				return err
+			}
+		case "fullScan":
+			err = (*jsonBool)(&sm.Key.FullScan).decodeJSON(iter.Value())
+			if err != nil {
+				return err
+			}
+		default:
+			if buildutil.CrdbTestBuild && panicOnUnknownFieldsForTest {
+				panic(errors.AssertionFailedf("unknown field: %s", iter.Key()))
+			}
+		}
+	}
+
+	return nil
+}
+
+func (sm *stmtStatsMetadata) decodeFlagsOnlyJSON(js json.JSON) (err error) {
+	iter, err := js.ObjectIter()
+	if err != nil {
+		return err
+	}
+
+	var fieldName string
+	defer func() {
+		if err != nil {
+			err = errors.Wrapf(err, "decoding field %s", fieldName)
+		}
+	}()
+
+	for ok := iter.Next(); ok; ok = iter.Next() {
+		switch iter.Key() {
+		case "db":
+			err = (*jsonString)(&sm.Key.Database).decodeJSON(iter.Value())
+			if err != nil {
+				return err
+			}
+		case "distsql":
+			err = (*jsonBool)(&sm.Key.DistSQL).decodeJSON(iter.Value())
+			if err != nil {
+				return err
+			}
+		case "implicitTxn":
+			err = (*jsonBool)(&sm.Key.ImplicitTxn).decodeJSON(iter.Value())
+			if err != nil {
+				return err
+			}
+		case "vec":
+			err = (*jsonBool)(&sm.Key.Vec).decodeJSON(iter.Value())
+			if err != nil {
+				return err
+			}
+		case "fullScan":
+			err = (*jsonBool)(&sm.Key.FullScan).decodeJSON(iter.Value())
+			if err != nil {
+				return err
+			}
+		default:
+			if buildutil.CrdbTestBuild && panicOnUnknownFieldsForTest {
+				panic(errors.AssertionFailedf("unknown field: %s", iter.Key()))
+			}
+		}
+	}
+
+	return nil
+}
+
+func (am *aggregatedMetadata) decodeAggregatedFieldsOnlyJSON(js json.JSON) (err error) {
+	iter, err := js.ObjectIter()
+	if err != nil {
+		return err
+	}
+
+	var fieldName string
+	defer func() {
+		if err != nil {
+			err = errors.Wrapf(err, "decoding field %s", fieldName)
+		}
+	}()
+
+	for ok := iter.Next(); ok; ok = iter.Next() {
+		switch iter.Key() {
+		case "db":
+			err = (*stringArray)(&am.Databases).decodeJSON(iter.Value())
+			if err != nil {
+				return err
+			}
+		case "appNames":
+			err = (*stringArray)(&am.AppNames).decodeJSON(iter.Value())
+			if err != nil {
+				return err
+			}
+		case "distSQLCount":
+			err = (*jsonInt)(&am.DistSQLCount).decodeJSON(iter.Value())
+			if err != nil {
+				return err
+			}
+		case "fullScanCount":
+			err = (*jsonInt)(&am.FullScanCount).decodeJSON(iter.Value())
+			if err != nil {
+				return err
+			}
+		case "implicitTxn":
+			err = (*jsonBool)(&am.ImplicitTxn).decodeJSON(iter.Value())
+			if err != nil {
+				return err
+			}
+		case "vecCount":
+			err = (*jsonInt)(&am.VecCount).decodeJSON(iter.Value())
+			if err != nil {
+				return err
+			}
+		case "totalCount":
+			err = (*jsonInt)(&am.TotalCount).decodeJSON(iter.Value())
+			if err != nil {
+				return err
+			}
+		default:
+			if buildutil.CrdbTestBuild && panicOnUnknownFieldsForTest {
+				panic(errors.AssertionFailedf("unknown field: %s", iter.Key()))
+			}
+		}
+	}
+
+	return nil
 }
 
 // DecodeAggregatedMetadataAggregatedFieldsOnlyJSON decodes the 'aggregated metadata' represented by
@@ -63,7 +347,7 @@ func DecodeAggregatedMetadataJSON(
 func DecodeAggregatedMetadataAggregatedFieldsOnlyJSON(
 	metadata json.JSON, result *appstatspb.AggregatedStatementMetadata,
 ) error {
-	return (*aggregatedMetadata)(result).jsonAggregatedFields().decodeJSON(metadata)
+	return (*aggregatedMetadata)(result).decodeAggregatedFieldsOnlyJSON(metadata)
 }
 
 // DecodeStmtStatsStatisticsJSON decodes the 'statistics' field and the
@@ -71,8 +355,44 @@ func DecodeAggregatedMetadataAggregatedFieldsOnlyJSON(
 // appstatspb.StatementStatistics.
 func DecodeStmtStatsStatisticsJSON(
 	jsonVal json.JSON, result *appstatspb.StatementStatistics,
-) error {
-	return (*stmtStats)(result).decodeJSON(jsonVal)
+) (err error) {
+	iter, err := jsonVal.ObjectIter()
+	if err != nil {
+		return err
+	}
+
+	var fieldName string
+	defer func() {
+		if err != nil {
+			err = errors.Wrapf(err, "decoding field %s", fieldName)
+		}
+	}()
+
+	for ok := iter.Next(); ok; ok = iter.Next() {
+		switch iter.Key() {
+		case "statistics":
+			err = (*innerStmtStats)(result).decodeJSON(iter.Value())
+			if err != nil {
+				return err
+			}
+		case "execution_statistics":
+			err = (*execStats)(&result.ExecStats).decodeJSON(iter.Value())
+			if err != nil {
+				return err
+			}
+		case "index_recommendations":
+			err = (*stringArray)(&result.IndexRecommendations).decodeJSON(iter.Value())
+			if err != nil {
+				return err
+			}
+		default:
+			if buildutil.CrdbTestBuild && panicOnUnknownFieldsForTest {
+				panic(errors.AssertionFailedf("unknown field: %s", iter.Key()))
+			}
+		}
+	}
+
+	return nil
 }
 
 // JSONToExplainTreePlanNode decodes the JSON-formatted ExplainTreePlanNode