--- a/pkg/sql/sqlstats/persistedsqlstats/sqlstatsutil/json_impl.go
+++ b/pkg/sql/sqlstats/persistedsqlstats/sqlstatsutil/json_impl.go
@@ -51,13 +51,6 @@ var (
 
 type txnStats appstatspb.TransactionStatistics
 
-func (t *txnStats) jsonFields() jsonFields {
-	return jsonFields{
-		{"statistics", (*innerTxnStats)(t)},
-		{"execution_statistics", (*execStats)(&t.ExecStats)},
-	}
-}
-
 func (t *txnStats) decodeJSON(js json.JSON) error {
 	// Decode "statistics" field
 	if valJSON, err := js.FetchValKey("statistics"); err != nil {
@@ -96,14 +89,6 @@ func (t *txnStats) encodeJSON() (json.JSON, error) {
 }
 
 type stmtStats appstatspb.StatementStatistics
-
-func (s *stmtStats) jsonFields() jsonFields {
-	return jsonFields{
-		{"statistics", (*innerStmtStats)(s)},
-		{"execution_statistics", (*execStats)(&s.ExecStats)},
-		{"index_recommendations", (*stringArray)(&s.IndexRecommendations)},
-	}
-}
 
 func (s *stmtStats) decodeJSON(js json.JSON) error {
 	// Decode "statistics" field
@@ -158,59 +143,7 @@ func (s *stmtStats) encodeJSON() (json.JSON, error) {
 
 type stmtStatsMetadata appstatspb.CollectedStatementStatistics
 
-func (s *stmtStatsMetadata) jsonFields() jsonFields {
-	return jsonFields{
-		{"stmtType", (*jsonString)(&s.Stats.SQLType)},
-		{"query", (*jsonString)(&s.Key.Query)},
-		{"querySummary", (*jsonString)(&s.Key.QuerySummary)},
-		{"db", (*jsonString)(&s.Key.Database)},
-		{"distsql", (*jsonBool)(&s.Key.DistSQL)},
-		{"implicitTxn", (*jsonBool)(&s.Key.ImplicitTxn)},
-		{"vec", (*jsonBool)(&s.Key.Vec)},
-		{"fullScan", (*jsonBool)(&s.Key.FullScan)},
-	}
-}
-
-func (s *stmtStatsMetadata) jsonFlagsOnlyFields() jsonFields {
-	return jsonFields{
-		{"db", (*jsonString)(&s.Key.Database)},
-		{"distsql", (*jsonBool)(&s.Key.DistSQL)},
-		{"implicitTxn", (*jsonBool)(&s.Key.ImplicitTxn)},
-		{"vec", (*jsonBool)(&s.Key.Vec)},
-		{"fullScan", (*jsonBool)(&s.Key.FullScan)},
-	}
-}
-
 type aggregatedMetadata appstatspb.AggregatedStatementMetadata
-
-func (s *aggregatedMetadata) jsonFields() jsonFields {
-	return jsonFields{
-		{"db", (*stringArray)(&s.Databases)},
-		{"appNames", (*stringArray)(&s.AppNames)},
-		{"distSQLCount", (*jsonInt)(&s.DistSQLCount)},
-		{"fullScanCount", (*jsonInt)(&s.FullScanCount)},
-		{"implicitTxn", (*jsonBool)(&s.ImplicitTxn)},
-		{"query", (*jsonString)(&s.Query)},
-		{"formattedQuery", (*jsonString)(&s.FormattedQuery)},
-		{"querySummary", (*jsonString)(&s.QuerySummary)},
-		{"stmtType", (*jsonString)(&s.StmtType)},
-		{"vecCount", (*jsonInt)(&s.VecCount)},
-		{"totalCount", (*jsonInt)(&s.TotalCount)},
-		{"fingerprintID", (*jsonString)(&s.FingerprintID)},
-	}
-}
-
-func (s *aggregatedMetadata) jsonAggregatedFields() jsonFields {
-	return jsonFields{
-		{"db", (*stringArray)(&s.Databases)},
-		{"appNames", (*stringArray)(&s.AppNames)},
-		{"distSQLCount", (*jsonInt)(&s.DistSQLCount)},
-		{"fullScanCount", (*jsonInt)(&s.FullScanCount)},
-		{"implicitTxn", (*jsonBool)(&s.ImplicitTxn)},
-		{"vecCount", (*jsonInt)(&s.VecCount)},
-		{"totalCount", (*jsonInt)(&s.TotalCount)},
-	}
-}
 
 type int64Array []int64
 
@@ -365,21 +298,6 @@ func (s *stmtFingerprintID) encodeJSON() (json.JSON, error) {
 }
 
 type innerTxnStats appstatspb.TransactionStatistics
-
-func (t *innerTxnStats) jsonFields() jsonFields {
-	return jsonFields{
-		{"cnt", (*jsonInt)(&t.Count)},
-		{"maxRetries", (*jsonInt)(&t.MaxRetries)},
-		{"numRows", (*numericStats)(&t.NumRows)},
-		{"svcLat", (*numericStats)(&t.ServiceLat)},
-		{"retryLat", (*numericStats)(&t.RetryLat)},
-		{"commitLat", (*numericStats)(&t.CommitLat)},
-		{"idleLat", (*numericStats)(&t.IdleLat)},
-		{"bytesRead", (*numericStats)(&t.BytesRead)},
-		{"rowsRead", (*numericStats)(&t.RowsRead)},
-		{"rowsWritten", (*numericStats)(&t.RowsWritten)},
-	}
-}
 
 func (t *innerTxnStats) decodeJSON(js json.JSON) error {
 	// Decode "cnt" field
@@ -538,36 +456,6 @@ func (t *innerTxnStats) encodeJSON() (json.JSON, error) {
 }
 
 type innerStmtStats appstatspb.StatementStatistics
-
-func (s *innerStmtStats) jsonFields() jsonFields {
-	return jsonFields{
-		{"cnt", (*jsonInt)(&s.Count)},
-		{"firstAttemptCnt", (*jsonInt)(&s.FirstAttemptCount)},
-		{"maxRetries", (*jsonInt)(&s.MaxRetries)},
-		{"lastExecAt", (*jsonTime)(&s.LastExecTimestamp)},
-		{"numRows", (*numericStats)(&s.NumRows)},
-		{"idleLat", (*numericStats)(&s.IdleLat)},
-		{"parseLat", (*numericStats)(&s.ParseLat)},
-		{"planLat", (*numericStats)(&s.PlanLat)},
-		{"runLat", (*numericStats)(&s.RunLat)},
-		{"svcLat", (*numericStats)(&s.ServiceLat)},
-		{"ovhLat", (*numericStats)(&s.OverheadLat)},
-		{"bytesRead", (*numericStats)(&s.BytesRead)},
-		{"rowsRead", (*numericStats)(&s.RowsRead)},
-		{"rowsWritten", (*numericStats)(&s.RowsWritten)},
-		{"nodes", (*int64Array)(&s.Nodes)},
-		{"kvNodeIds", (*int32Array)(&s.KVNodeIDs)},
-		{"regions", (*stringArray)(&s.Regions)},
-		{"usedFollowerRead", (*jsonBool)(&s.UsedFollowerRead)},
-		{"planGists", (*stringArray)(&s.PlanGists)},
-		{"indexes", (*stringArray)(&s.Indexes)},
-		{"latencyInfo", (*latencyInfo)(&s.LatencyInfo)},
-		{"lastErrorCode", (*jsonString)(&s.LastErrorCode)},
-		{"failureCount", (*jsonInt)(&s.FailureCount)},
-		{"genericCount", (*jsonInt)(&s.GenericCount)},
-		{"sqlType", (*jsonString)(&s.SQLType)},
-	}
-}
 
 func (s *innerStmtStats) decodeJSON(js json.JSON) (err error) {
 	var fieldName string
@@ -852,19 +740,6 @@ func (s *innerStmtStats) encodeJSON() (json.JSON, error) {
 
 type execStats appstatspb.ExecStats
 
-func (e *execStats) jsonFields() jsonFields {
-	return jsonFields{
-		{"cnt", (*jsonInt)(&e.Count)},
-		{"networkBytes", (*numericStats)(&e.NetworkBytes)},
-		{"maxMemUsage", (*numericStats)(&e.MaxMemUsage)},
-		{"contentionTime", (*numericStats)(&e.ContentionTime)},
-		{"networkMsgs", (*numericStats)(&e.NetworkMessages)},
-		{"maxDiskUsage", (*numericStats)(&e.MaxDiskUsage)},
-		{"cpuSQLNanos", (*numericStats)(&e.CPUSQLNanos)},
-		{"mvccIteratorStats", (*iteratorStats)(&e.MVCCIteratorStats)},
-	}
-}
-
 func (e *execStats) decodeJSON(js json.JSON) (err error) {
 	var fieldName string
 	defer func() {
@@ -978,24 +853,6 @@ func (e *execStats) encodeJSON() (json.JSON, error) {
 }
 
 type iteratorStats appstatspb.MVCCIteratorStats
-
-func (e *iteratorStats) jsonFields() jsonFields {
-	return jsonFields{
-		{"stepCount", (*numericStats)(&e.StepCount)},
-		{"stepCountInternal", (*numericStats)(&e.StepCountInternal)},
-		{"seekCount", (*numericStats)(&e.SeekCount)},
-		{"seekCountInternal", (*numericStats)(&e.SeekCountInternal)},
-		{"blockBytes", (*numericStats)(&e.BlockBytes)},
-		{"blockBytesInCache", (*numericStats)(&e.BlockBytesInCache)},
-		{"keyBytes", (*numericStats)(&e.KeyBytes)},
-		{"valueBytes", (*numericStats)(&e.ValueBytes)},
-		{"pointCount", (*numericStats)(&e.PointCount)},
-		{"pointsCoveredByRangeTombstones", (*numericStats)(&e.PointsCoveredByRangeTombstones)},
-		{"rangeKeyCount", (*numericStats)(&e.RangeKeyCount)},
-		{"rangeKeyContainedPoints", (*numericStats)(&e.RangeKeyContainedPoints)},
-		{"rangeKeySkippedPoints", (*numericStats)(&e.RangeKeySkippedPoints)},
-	}
-}
 
 func (e *iteratorStats) decodeJSON(js json.JSON) (err error) {
 	var fieldName string
@@ -1161,13 +1018,6 @@ func (e *iteratorStats) encodeJSON() (json.JSON, error) {
 
 type numericStats appstatspb.NumericStat
 
-func (n *numericStats) jsonFields() jsonFields {
-	return jsonFields{
-		{"mean", (*jsonFloat)(&n.Mean)},
-		{"sqDiff", (*jsonFloat)(&n.SquaredDiffs)},
-	}
-}
-
 func (n *numericStats) decodeJSON(js json.JSON) error {
 	// Decode "mean" field
 	if valJSON, err := js.FetchValKey("mean"); err != nil {
@@ -1228,13 +1078,6 @@ func (n *numericStats) encodeJSONWithBuilder(
 }
 
 type latencyInfo appstatspb.LatencyInfo
-
-func (l *latencyInfo) jsonFields() jsonFields {
-	return jsonFields{
-		{"min", (*jsonFloat)(&l.Min)},
-		{"max", (*jsonFloat)(&l.Max)},
-	}
-}
 
 func (l *latencyInfo) decodeJSON(js json.JSON) error {
 	// Decode "min" field