## sqlstats: inline jsonFields and use iterator during 

Previously, we would look up individual fields in the JSON objects
when decoding, this would result in unnecessary iteration through
the object. We were also constructing a `jsonFields` on every decode
request.

```
❯ benchdiff ./pkg/sql/sqlstats/persistedsqlstats/sqlstatsutil/ --old https://github.com/cockroachdb/cockroach/commit/0e9251dba2fd91e153974e2701f73c93871f9d91 -b -r 'SQLStatsJson' -c 10 --preview
old:  https://github.com/cockroachdb/cockroach/commit/0e9251dba2fd91e153974e2701f73c93871f9d91 Merge https://github.com/cockroachdb/cockroach/pull/150535 https://github.com/cockroachdb/cockroach/pull/150771 https://github.com/cockroachdb/cockroach/pull/150947 https://github.com/cockroachdb/cockroach/pull/150949
new:  e768737 sqlstats: inline jsonFields during encode
args: benchdiff "./pkg/sql/sqlstats/persistedsqlstats/sqlstatsutil/" "--old" "0e9251db" "-b" "-r" "SQLStatsJson" "-c" "10" "--preview"

name                                         old time/op    new time/op    delta
SQLStatsJson/statement_metadata/decoding-10     406ns ± 1%     117ns ± 2%   -71.10%  (p=0.000 n=9+9)
SQLStatsJson/statement_stats/decoding-10       6.62µs ± 1%    4.41µs ± 0%   -33.36%  (p=0.000 n=9+10)
SQLStatsJson/transaction_stats/decoding-10     4.96µs ± 2%    3.67µs ± 0%   -26.08%  (p=0.000 n=10+9)

name                                         old speed      new speed      delta
SQLStatsJson/statement_metadata/decoding-10   986MB/s ± 1%  3412MB/s ± 2%  +245.96%  (p=0.000 n=9+9)
SQLStatsJson/statement_stats/decoding-10     1.11GB/s ± 1%  1.88GB/s ± 0%   +69.66%  (p=0.000 n=9+10)
SQLStatsJson/transaction_stats/decoding-10   1.14GB/s ± 2%  1.54GB/s ± 0%   +35.27%  (p=0.000 n=10+9)

name                                         old alloc/op   new alloc/op   delta
SQLStatsJson/statement_metadata/decoding-10      432B ± 0%       80B ± 0%   -81.48%  (p=0.000 n=10+10)
SQLStatsJson/statement_stats/decoding-10       3.54kB ± 0%    2.38kB ± 0%   -32.58%  (p=0.000 n=10+10)
SQLStatsJson/transaction_stats/decoding-10     2.56kB ± 0%    1.89kB ± 0%   -26.25%  (p=0.000 n=10+10)

name                                         old allocs/op  new allocs/op  delta
SQLStatsJson/transaction_stats/decoding-10       59.0 ± 0%      59.0 ± 0%      ~     (all equal)
SQLStatsJson/statement_metadata/decoding-10      4.00 ± 0%      4.00 ± 0%      ~     (all equal)
SQLStatsJson/statement_stats/decoding-10         75.0 ± 0%      78.0 ± 0%    +4.00%  (p=0.000 n=10+10)  
```

Release note: None

----

## sqlstats: inline jsonFields during encode

Previously, we would construct instances of `jsonFields` during
encoding which was unnecessary. This change unrolls those operations
and also uses a shared JSON builder object to reduce allocations.

```
❯ benchdiff ./pkg/sql/sqlstats/persistedsqlstats/sqlstatsutil/ --old https://github.com/cockroachdb/cockroach/commit/0e9251dba2fd91e153974e2701f73c93871f9d91 -b -r 'SQLStatsJson' -c 10 --preview
old:  https://github.com/cockroachdb/cockroach/commit/0e9251dba2fd91e153974e2701f73c93871f9d91 Merge https://github.com/cockroachdb/cockroach/pull/150535 https://github.com/cockroachdb/cockroach/pull/150771 https://github.com/cockroachdb/cockroach/pull/150947 https://github.com/cockroachdb/cockroach/pull/150949
new:  e768737 sqlstats: inline jsonFields during encode
args: benchdiff "./pkg/sql/sqlstats/persistedsqlstats/sqlstatsutil/" "--old" "0e9251db" "-b" "-r" "SQLStatsJson" "-c" "10" "--preview"

name                                         old time/op    new time/op    delta
SQLStatsJson/statement_metadata/encoding-10     738ns ± 1%     564ns ± 1%   -23.51%  (p=0.000 n=9+9)
SQLStatsJson/transaction_stats/encoding-10     13.3µs ± 1%    12.3µs ± 1%    -7.31%  (p=0.000 n=9+10)
SQLStatsJson/statement_stats/encoding-10       16.5µs ± 1%    15.6µs ± 1%    -5.32%  (p=0.000 n=10+9)

name                                         old speed      new speed      delta
SQLStatsJson/statement_metadata/encoding-10  1.01GB/s ± 1%  1.32GB/s ± 1%   +30.75%  (p=0.000 n=9+9)
SQLStatsJson/transaction_stats/encoding-10   43.1MB/s ± 1%  46.5MB/s ± 1%    +7.89%  (p=0.000 n=9+10)
SQLStatsJson/statement_stats/encoding-10     45.0MB/s ± 1%  47.6MB/s ± 1%    +5.62%  (p=0.000 n=10+9)

name                                         old alloc/op   new alloc/op   delta
SQLStatsJson/statement_metadata/encoding-10    1.13kB ± 0%    0.74kB ± 0%   -34.04%  (p=0.000 n=10+10)
SQLStatsJson/transaction_stats/encoding-10     9.79kB ± 0%    8.14kB ± 0%   -16.83%  (p=0.000 n=10+10)
SQLStatsJson/statement_stats/encoding-10       12.5kB ± 0%    11.9kB ± 0%    -4.74%  (p=0.000 n=10+10)

name                                         old allocs/op  new allocs/op  delta
SQLStatsJson/transaction_stats/encoding-10        296 ± 0%       255 ± 0%   -13.85%  (p=0.000 n=10+10)
SQLStatsJson/statement_stats/encoding-10          337 ± 0%       293 ± 0%   -13.06%  (p=0.000 n=10+10)
SQLStatsJson/statement_metadata/encoding-10      11.0 ± 0%      10.0 ± 0%    -9.09%  (p=0.000 n=10+10)
```

Release note: None